requirements.txt
PyQt6>=6.7
edge-tts>=7.0
auto-editor>=27.0
Pillow>=10.4
numpy>=1.26
imageio-ffmpeg>=0.5
